DevFest25-WS/P4-Memory/pyproject.toml
[project]
name = "p4-memory"
version = "0.1.0"
description = "Part 4: Agent with Memory - The Adaptive Planner"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai",
    "python-dotenv",
]

